refactor(type): separate util function and constants types
