+++ src/shared/type/MessageType.js
@@ -1,0 +1,10 @@
+/**
+ * @typedef {Object} MessageType
+ * @type {{lint: string, ignore: string}}
+ */
+export default {
+    // Linting(also include Fix) Message Type
+    "lint": "lint",
+    // Ignoring Message Type
+    "ignore": "ignore"
+};
