+++ src/shared/type/SeverityLevel.js
@@ -1,0 +1,11 @@
+// LICENSE : MIT
+"use strict";
+/**
+ * @typedef {Object} SeverityLevel
+ */
+export default {
+    "none": 0,
+    "info": 0,
+    "warning": 1,
+    "error": 2
+};
